Better Grass Implementation
Grass is only shaded on sufficiently flat areas. Similar condition is applied to sand but the additional condition of being within 5% of seaLevel is present. The ability of grass to grow is determined by the grassThreshold variable.

Implemented primarily in DrawHeights class.

Currently colours of sand and other non-grass land needs to be fixed.
Sand often has lines of grass within it. Erosion would likely fix this particular issue.